requirements.txt
unreal-script-editor